Refactored classes in the dto directory
